feat(screenshots): Add `captureBeyondViewport` option to screenshot
The default behavior for Playwright in chromium is to temporarily resize
the viewport before taking a screenshot. This happens both for
`page.screenshot({ fullPage: true })` and `elementHandle.screenshot()`.
Without it, screenshots would be cut at the viewport size. In general,
this behavior/workaround is okay. But when you start using CSS viewport
units like vh and vw, screenshots will look wrong. Consider the
following example:

```
<main style="min-height: 2000px">
  <div style="height: 50vh"></div>
</main>
```

For a viewport of size 800px, the div will render at 400px. But a
Playwright screenshot of a this page will show the div at 1000px. Of
course, one could argue that you should probably use vh heights together
with a fixed max-height or vmax or similar (with the reasoning usually
being something like "people shouldn't assume viewports are always
small") but the fact is that there are tons of websites out there that
use viewport units and they all sort of look wrong in Playwright
screenshots.

At happo.io, we've been using Puppeteer for an extended period of time,
and I've noticed that they don't have the same problem. Viewport units
work well even when content doesn't fit inside the viewport. After doing
some digging, I found that Puppeteer passes the `captureBeyondViewport`
flag in the call to `captureScreenshot()`.
https://github.com/puppeteer/puppeteer/blob/4d9dc8c0e613f22d4cdf237e8bd0b0da3c588edb/src/common/Page.ts#L2778
https://chromedevtools.github.io/devtools-protocol/tot/Page/#method-captureScreenshot

The `captureBeyondViewport` option is experiemental, so it's probably
best not to lock ourselves into the behavior by making it the default.
That's why this commit will add a new option to the screenshot methods
(page and elementHandle). That way, people can opt in to the
experimental behavior and people who are happy with the viewport resize
workaround can stay with the default.

This is a chromium only option. Other browsers won't be affected by this
change (they are also already doing "the right thing" for screenshots
with viewport units).