browser(firefox): fix HTTP->HTTPS fallback for refused connection
In this case, there's no protocol version that we can extract for
nsIHttpChannel.

The code that does the redirect is here: https://github.com/mozilla/gecko-dev/blob/7f3d9fce4154e71ec1af78eb6bcded0db11af08d/docshell/base/nsDocShell.cpp#L6079-L6095

To trigger this codepath, we'd need to run test inside a special
docker container that has https server running on the 443 port. We lack
infrastructure for this kind of tests (but it'll be cool to have it).

References #11118