fix(test runner): use a separate cache when baseUrl is set
Currently, with baseUrl, we end up writing the map file but not the
source file. This leads to multiple workers concurrently writing and
reading the same source map, instead of it being written just once
by the runner process.

This change uses a unique cache directory when baseUrl is present,
cleaned up after the run.