Revert "chore: best-effort cleanup for output folders that are mounted (#12300)"
This reverts commit e5c9d1e39fd0e4bb30e25d69adba2cc7ae3f0071.

Reason for revert: turns out this fix results in a 5-second delay
when starting tests in docker, with `test-results` folder being
a non-removable mount.

The reason for the delay is the `maxBusyTries` option that we
supply by default to rimraf when trying to remove the folder.

While this option might come handy when removing temporary
browser profile folder, it doesn't serve us well in this particular
usecase.