Revert "cherry-pick(#12300): chore: best-effort cleanup for output folders that are mounted (#12318)"
This reverts commit 0a1c1dad6769ed23ca808936c1d43645a6778a45.

Reason for revert: turns out this fix results in a 5-second delay
when starting tests in docker, with `test-results` folder being
a non-removable mount.

The reason for the delay is the `maxBusyTries` option that we
supply by default to rimraf when trying to remove the folder.

While this option might come handy when removing temporary
browser profile folder, it doesn't serve us well in this particular
usecase.

References #12106