Fixing example code in `accessibility` class
The example code for finding the focused node doesn't work as described. While this is not a particularly big deal, it should probably do what it says it does!

This patch updates the existing JS/Python examples to work as intended, but does't add any new code for missing languages.