docs: automatically test examples
This change intends to help us make our docs and examples:

1. more `@playwright/test` centric since many users consume the APIs via
   test runner
2. self-contained: many of our docs show example API usage, but don't
   show the corresponding DOM that the API is being exercised against
3. working out of the box: there is nothing more frustrating than trying
   an example only to get an error

This introduces:

```
$ npm run dtest
```

which will run all (annotated) code samples from the docs.

See the diffs of the `*.md` files for examples.

The current annotation/generation assumes the example is completely
self-contained, but since importing `test`/`expect` into each example
can get verbose (for both doc contributors and users/readers), we might
consider having that part implied and just including it as part of the
generation. Currently, runnable blocks are annotated with `RUNNABLE`,
so maybe we could introduce variations of this that indicate to the
generation step whether the example is self-container or requires
auto-adding of the common imports.