feat(test-runner): add proof-of-concept WebServerPlugin
I'm starting to think through what hooks we want to expose in plugins
and how the new API looks.

This is a quick hack to play around a bit. We implement our current
webServer config as a Plugin instead. This allows you to have multiple
WebServers! (Or any other global processes that the runner will setup
and teardown.)

This current patch is not intended to land. It's me learning the
internals of the Runner and starting to see where we can hook into and
think through some APIs. I have quite a few ideas of what could be
customized (see related issue), but we don't necessarily want to do that
all today (e.g. influencing scheduler/placement decisions), but we
should design something that could be extended to support deeper hooking
and customization in the future.

The next interesting progression of this experiment is to allow a Plugin
to mofify the Config before we run.

This way, a plugin could modify multiple settings at once (e.g. add a
custom reporter, adjust worker counts, connection options, etc.).

Relates #13362
Relates #8206