feat: add serviceWorkerPolicy option
Anecdotally, it's common for users to want to disable Service Workers
while testing, so this provides a simple context option to do so.

It also prepares us to land #14321 which requires us to have a point for
folks to enable seeing Service Worker Requests as events (and allow for
routing).