fix(test runner): disallow test.use() in require'd files
When `test.use()` is called not directly in a test file,
but rahter in a require'd file, it leads to subtle bugs.
Only the first test file that requires the helper with `test.use()`
is going to get the use options. This leads to confusion
and bugs.

Here we check the stack to detect the situation where
`test.use()` is not called directly in the test file and
print a nice error message.