chore: vendor proper-lockfile dependency
This patch vendors the https://github.com/moxystudio/node-proper-lockfile
project we rely to manage Playwright Registry.

The reason to vender is the following upstream issue that
didn't get resolved in almost a month: https://github.com/moxystudio/node-proper-lockfile/issues/111

Follow-up will apply the fix for the issue to the vendored file.

NOTE: this patch also inlines all code in a single file.

References #19418