fix: better formatting for sparse arrays
Right now arrays preview yields all array elements. In case
of a sparse array with a single element on index 10000000,
this results in a large string that OOM Node.js.

This patch changes pretty-printing. For example:

```ts
// Given this array
const a = [];
a[10] = 1;
// Before this patch, pretty printing will yield:
"[,,,,,,,,1]"
// With this patch, pretty printing yields:
"[<9 empty items>,1]"
```

The new array pretty-printing is equal to what Node.js CLI
does to render sparse arrays.

Fixes #20347