fix(reporter-junit): Allow users to customize the class name for results in JUNIT reporter
Introduce a `classNameTemplate` property into the junit configuration who's value is a function that takes in a TestClass object and returns a string. By using properties of the TestClass Object Users can set the reported class name in the test junit results to whatever they want.
Updated doc with information and an example of this.

Fixes #20380