fix(trace viewer): prefer latest resource with the same url
When rendering snapshot, disregard earlier resources with the same url,
because it's most likely that the latest one was used for rendering.

An example would be reloading the page before the stylesheet has
finised loading. In this case, the stylesheet will be requested twice,
and the second copy that was not aborted should be used for the
snapshot.