feat(crx): playwright-crx package
Playwright chromium-flavored implementation compiled to run in a chrome
extension service worker.

The core of this integration is a ConnectionTransport implementation
(crxTransport.ts) that makes use of chrome.debugger.sendCommand and
chrome.debugger.onEvent for CDP instructions.

@see https://developer.chrome.com/docs/extensions/reference/debugger/

For testing, the approach is to run each test function in a chrome
extension service worker.
A test chrome extension is provided at tests/crx/test-extension, and
then each test function is serialized with required fixtures and is ran
there.
This is not yet a perfect solution, but provides a good test coverage
for next developments.

To run tests against playwright-crx, there's a npm xtest script