fix(junit reporter): Properly report actual errors in the junit failure message property
Currently junit test failure are reported with the failing test file's name, failure location, and test name in the message property, while the actual failure information is stored in the CDATA text. This can break tools that look at the error message to do failure analysis and deduplication as they will not have the actual failure cause.

This change makes it so we report all unique errors hit by the tests in the xml failure entry's message property.

#12944