fix(trace viewer): retain `currentSrc` of all images
When `<source>` or `srcset=` are involved, the actual image src
is determinted at runtime based on factors like `devicePixelRatio`
and media queries that depend on width/height.

Since these factors may differ in the Trace Viewer itself,
we should preserve the `currentSrc`, use it as an actual `src`,
and disable various `<source>` and `srcset=`.