cherry-pick(#27162): fix(tracing): bump trace version to V5, migrate V4 traces to consoleMessage.args (#27193)
This moves the fix in #27095 from `modernize` to `appendEvent`. The
reason is that `trace V4` is used both for older traces that do not have
`consoleMessage.args` and the new ones with `args`. Since we do not call
`modernize` for traces of the same version, the original fix does not
help in this case.

Fixes #27144.