fix(chromium): avoid premature continue of redirects
This has recently regressed in https://github.com/microsoft/playwright/pull/27429.
We now continue requests that are paused for the second time.

However, redirects share `networkId` with the original request, so we
may confuse paused redirect with a second pause for the original request.

This is covered by the flaky test
`page-route.spec.ts:392 > should work with redirects for subresources`